requirements.txt
# 猫狗分类任务依赖库
# 安装命令: pip install -r requirements.txt

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 数据处理和可视化
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
tensorboard>=2.13.0
seaborn>=0.12.0
scikit-learn>=1.2.0

# ONNX相关
onnx>=1.14.0
onnxruntime>=1.15.0
onnxscript>=0.1.0
